refactor result builder as create function in quantstudio
