+++ tests/parsers/appbio_quantstudio/appbio_quantstudio_builders_test.py
@@ -10,7 +10,6 @@ from allotropy.allotrope.allotrope import AllotropeConversionError
 from allotropy.allotrope.models.pcr_benchling_2023_09_qpcr import ExperimentType
 from allotropy.parsers.appbio_quantstudio.appbio_quantstudio_builders import (
     DataBuilder,
-    ResultsBuilder,
 )
 from allotropy.parsers.appbio_quantstudio.appbio_quantstudio_structure import (
     Data,
@@ -165,9 +164,7 @@ def test_results_builder() -> None:
         quencher_dye_setting=None,
         sample_role_type="PC_ALLELE_1",
     )
-    result = ResultsBuilder.build(
-        data, well_item, ExperimentType.genotyping_qPCR_experiment
-    )
+    result = Result.create(data, well_item, ExperimentType.genotyping_qPCR_experiment)
     assert isinstance(result, Result)
     assert result.cycle_threshold_value_setting == 0.219
     assert result.cycle_threshold_result is None
